Resolving code reviews
