+++ frontend/scripts/register-typescript-loader.mjs
@@ -1,0 +1,4 @@
+import { register } from "node:module";
+import { pathToFileURL } from "node:url";
+
+register("./scripts/typescript-loader.mjs", pathToFileURL("./"));
