+++ frontend/scripts/typescript-loader.mjs
@@ -1,0 +1,20 @@
+import { readFile } from "node:fs/promises";
+import ts from "typescript";
+
+export async function load(url, context, nextLoad) {
+  if (!url.endsWith(".ts")) return nextLoad(url, context);
+
+  const source = await readFile(new URL(url), "utf8");
+  const transpiled = ts.transpileModule(source, {
+    compilerOptions: {
+      module: ts.ModuleKind.ESNext,
+      target: ts.ScriptTarget.ES2020,
+    },
+  });
+
+  return {
+    format: "module",
+    shortCircuit: true,
+    source: transpiled.outputText,
+  };
+}
