+++ frontend/src/features/authoring/util.ts
@@ -158,7 +158,6 @@ export function filterComponent(component: Record<string, unknown>) {
     fill: component.fill,
     preserveAspectRatio: component.preserveAspectRatio,
     href: component.href,
-    opacity: component.opacity,
   };
   return filtered;
 }
